requirements.txt
numpy
pillow
opencv-python
mediapipe==0.10.14
rembg
onnxruntime
